doc: add more comments
